refactor URI helpers
