+++ mlflow/utils/__init__.py
@@ -1,12 +1,7 @@
 from sys import version_info
 
-from six.moves import urllib
-
-from mlflow.exceptions import MlflowException
-from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
-
-from mlflow.store.db.db_types import DATABASE_ENGINES
 from mlflow.utils.annotations import deprecated, experimental, keyword_only
+from mlflow.utils.uri import extract_db_type_from_uri
 from mlflow.utils.validation import _validate_db_type_string
 
 PYTHON_VERSION = "{major}.{minor}.{micro}".format(major=version_info.major,
@@ -14,27 +9,6 @@ PYTHON_VERSION = "{major}.{minor}.{micro}".format(major=version_info.major,
                                                   micro=version_info.micro)
 _INVALID_DB_URI_MSG = "Please refer to https://mlflow.org/docs/latest/tracking.html#storage for " \
                       "format specifications."
-
-
-def extract_db_type_from_uri(db_uri):
-    """
-    Parse the specified DB URI to extract the database type. Confirm the database type is
-    supported. If a driver is specified, confirm it passes a plausible regex.
-    """
-    scheme = urllib.parse.urlparse(db_uri).scheme
-    scheme_plus_count = scheme.count('+')
-
-    if scheme_plus_count == 0:
-        db_type = scheme
-    elif scheme_plus_count == 1:
-        db_type, _ = scheme.split('+')
-    else:
-        error_msg = "Invalid database URI: '%s'. %s" % (db_uri, _INVALID_DB_URI_MSG)
-        raise MlflowException(error_msg, INVALID_PARAMETER_VALUE)
-
-    _validate_db_type_string(db_type)
-
-    return db_type
 
 
 def get_major_minor_py_version(py_version):
@@ -69,11 +43,3 @@ def get_unique_resource_id(max_length=None):
     if max_length is not None:
         unique_id = unique_id[:int(max_length)]
     return unique_id
-
-
-def get_uri_scheme(uri_or_path):
-    scheme = urllib.parse.urlparse(uri_or_path).scheme
-    if any([scheme.lower().startswith(db) for db in DATABASE_ENGINES]):
-        return extract_db_type_from_uri(uri_or_path)
-    else:
-        return scheme
