+++ tests/tracking/test_utils.py
@@ -10,8 +10,7 @@ from mlflow.store.tracking.rest_store import RestStore
 from mlflow.store.tracking.sqlalchemy_store import SqlAlchemyStore
 from mlflow.tracking.registry import TrackingStoreRegistry
 from mlflow.tracking.utils import _get_store, _TRACKING_URI_ENV_VAR, _TRACKING_USERNAME_ENV_VAR, \
-    _TRACKING_PASSWORD_ENV_VAR, _TRACKING_TOKEN_ENV_VAR, _TRACKING_INSECURE_TLS_ENV_VAR, \
-    get_db_profile_from_uri
+    _TRACKING_PASSWORD_ENV_VAR, _TRACKING_TOKEN_ENV_VAR, _TRACKING_INSECURE_TLS_ENV_VAR
 
 
 def test_get_store_file_store(tmp_wkdir):
@@ -280,7 +279,3 @@ def test_get_store_for_unregistered_scheme():
     with pytest.raises(mlflow.exceptions.MlflowException,
                        match="Unexpected URI scheme"):
         tracking_store.get_store("unknown-scheme://")
-
-
-def test_get_db_profile_from_uri_casing():
-    assert get_db_profile_from_uri('databricks://aAbB') == 'aAbB'
