+++ tests/utils/test_utils.py
@@ -1,8 +1,6 @@
 import pytest
 
-from mlflow.exceptions import MlflowException
-from mlflow.store.db.db_types import DATABASE_ENGINES
-from mlflow.utils import get_unique_resource_id, extract_db_type_from_uri, get_uri_scheme
+from mlflow.utils import get_unique_resource_id
 
 
 def test_get_unique_resource_id_respects_max_length():
@@ -17,18 +15,3 @@ def test_get_unique_resource_id_with_invalid_max_length_throws_exception():
 
     with pytest.raises(ValueError):
         get_unique_resource_id(max_length=0)
-
-
-def test_extract_db_type_from_uri():
-    uri = "{}://username:password@host:port/database"
-    for legit_db in DATABASE_ENGINES:
-        assert legit_db == extract_db_type_from_uri(uri.format(legit_db))
-        assert legit_db == get_uri_scheme(uri.format(legit_db))
-
-        with_driver = legit_db + "+driver-string"
-        assert legit_db == extract_db_type_from_uri(uri.format(with_driver))
-        assert legit_db == get_uri_scheme(uri.format(with_driver))
-
-    for unsupported_db in ["a", "aa", "sql"]:
-        with pytest.raises(MlflowException):
-            extract_db_type_from_uri(unsupported_db)
